chore: Add methods for searching projects and lists of tasks by keyword
This commit adds new methods to the ProjectManager and ListToDoManager classes for searching projects and lists of tasks by keyword. The ProjectManager class now includes a method called containsKeyword() that checks if a project's title, description, or type contains the specified keyword. The ListToDoManager class includes containsKeyword() and searchProjectsByKeyword() methods that perform similar searches on lists of tasks. These methods will improve the functionality of the application by allowing users to easily find relevant projects and tasks based on keywords.